(feat) moved attendance related queries to attendance model
