+++ src/server/db/QueryModels/AttendanceModel.js
@@ -1,6 +1,7 @@
 import AttendanceQueries from '../QuerySelectors/AttendanceQueries';
 import Promise from 'bluebird';
 import db from '../index.js';
+import { sendAbsentEmails } from '../../mailgun/mailgunHelpers';
 
 Promise.promisifyAll(db);
 
@@ -8,6 +9,16 @@ export default class AttendanceModel extends AttendanceQueries {
 
   constructor() {
     super();
+  }
+
+  async getStudentRecord(email) {
+    const queryString = super.studentAttendance(email);
+    return await db.queryAsync(queryString);
+  }
+
+  async getAllRecords() {
+    const queryString = super.allAttendance();
+    return await db.queryAsync(queryString);
   }
 
   async storeRecords(classes) {
@@ -20,7 +31,6 @@ export default class AttendanceModel extends AttendanceQueries {
       if (userDate[0].length) {
         let existingDay = new Date(userDate[0][0].date).getDay();
         let currentDay = new Date().getDay();
-        console.log(existingDay, currentDay)
         if (existingDay !== currentDay) {
           db.queryAsync(insertQuery);  
         }
@@ -30,4 +40,17 @@ export default class AttendanceModel extends AttendanceQueries {
     });
   }
 
+  async emailLateStudents() {
+    const getLateUsersQuery = super.getPendingUsers();
+    const getAllLateEmailsQuery = super.getAllLateUserEmails();
+    const [users] = await db.queryAsync(getLateUsersQuery);
+    users.forEach(async (user) => {
+      let lateQuery = super.pendingToAbsent(user.users_id);
+      await db.queryAsync(lateQuery);
+    });
+    const [emails] = await db.queryAsync(getAllLateEmailsQuery);
+    await sendAbsentEmails(emails);
+  }
 }
+
+
